ETL.wearther/requirements.txt
# Librairies utilitaires
psycopg2-binary==2.9.9
requests==2.31.0
pendulum==2.1.2

# Providers Airflow (ajoute ceux nécessaires à tes DAGs)
apache-airflow-providers-postgres
apache-airflow-providers-http
apache-airflow-providers-common-sql